perf(notebook-cloud): lazy load notebook route
